+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   * Created `DummySwitch` to be able to test Switch-Commit without a HardwareSwitch
   * Backend now uses `AutoDetectSwitch` to enable the future use of other Switch brands without changeing any backend-code
   * added `commitSystem` UnitTest to validate the configuration of HWswitches and dynamic created services on future (core) changes
+  * added prefetcher to download docker images, used by dynaimc created containers, in the backgroud before they are needed
 
 ## v0.8.0